requirements.txt
minio
asyncio-nats-client